feat!: more re-organization and name changes
