+++ src/Daybreak/Common/Features/Tiles/WaterFoliage/LilyPad.cs
